move checkPath method
