+++ fe/fe-core/src/main/java/org/apache/doris/analysis/LoadStmt.java
@@ -430,7 +430,7 @@ public class LoadStmt extends DdlStmt {
                     String location = brokerDesc.getFileLocation(dataDescription.getFilePaths().get(i));
                     dataDescription.getFilePaths().set(i, location);
                     dataDescription.getFilePaths().set(i,
-                            ExportStmt.checkPath(dataDescription.getFilePaths().get(i), brokerDesc.getStorageType()));
+                            StorageBackend.checkPath(dataDescription.getFilePaths().get(i), brokerDesc.getStorageType()));
                 }
             }
         }
